requirements.txt
altair==5.3.0
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
blinker==1.8.2
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==25.1
django-model-utils==4.5.1
django-notifications-hq==1.8.3
django-role-permissions==3.2.0
django-simple-history==3.8.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.7
Jinja2==3.1.4
jsonfield==3.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mssql-django==1.4
mysqlclient==2.2.4
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
prompt_toolkit==3.0.46
protobuf==5.27.1
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==1.7.1
pyodbc==5.1.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.5
referencing==0.35.1
reportlab==4.3.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.35
sqlparse==0.4.4
streamlit==1.36.0
swapper==1.3.0
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
watchdog==4.0.1
wcwidth==0.2.13
